auto-show when migration begins
